player health double?!?!?!?!

+++ zuul-framework/src/TWoT_A1/TWoT.java
@@ -843,7 +843,7 @@ public class TWoT implements Serializable{
      * 
      * @return 
      */
-    public double getPlayerHealth(){
+    public int getPlayerHealth(){
         return player.getHealth();
     }
     /**
