+++ zuul-framework/src/TWoT_test/TWoT.java
@@ -861,7 +861,7 @@ public class TWoT implements Serializable{
      * 
      * @return 
      */
-    public double getPlayerHealth(){
+    public int getPlayerHealth(){
         return player.getHealth();
     }
     /**
